Add new options, expand prohibited characters.
The new options are length, whitespace, and replace.
Length is how long the name should be truncated to, if it is too long.
Padding is subtracted from the length. Whitespace is a string that will
replace any whitespace characters after they have been stripped from the
beginning and end of the string. Multiple whitespace characters are
still combined into one. Replace is a string that will replace
prohibited characters rather than deleting them.